Flytt kode for modal-innhold inn i modal-komponent så data rendres riktig
